+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Services and managers
 Views
 
 - `FetchDelegatingAsyncImage` is added that allows you to write custom fetch function
+- `possible` is added to `BaseButtonGestureState`, and some `case`s are renamed
 - `HOrVStack` is renamed to `HVStack`
 
 Models